Project mavenized
